refactor : 피드백 반영 - 5
- SRP 준수를 위해 UserDatabase와 ArticleDatabase 분리
